add a new function: createBitmap
to convert latex string to a bitmap
